+++ packages/apps/graph/src/services/chainweb-node/mempool.ts
@@ -2,6 +2,7 @@ import type { Signer } from '@prisma/client';
 import { chainIds } from '@utils/chains';
 import { dotenv } from '@utils/dotenv';
 import https from 'https';
+import { networkConfig } from '../..';
 
 export class MempoolError extends Error {
   public mempoolError: any;
@@ -13,11 +14,13 @@ export class MempoolError extends Error {
 }
 
 export async function mempoolGetPending() {
+  const { networkId, apiVersion } = await networkConfig;
+
   return new Promise((resolve, reject) => {
     const options = {
       hostname: dotenv.MEMPOOL_HOSTNAME,
       port: dotenv.MEMPOOL_PORT,
-      path: '/chainweb/0.0/development/chain/1/mempool/getPending',
+      path: `/chainweb/${apiVersion}/${networkId}/chain/0/mempool/getPending`,
       method: 'POST',
       rejectUnauthorized: false, // This disables certificate verification
       headers: {
@@ -53,6 +56,7 @@ export async function mempoolGetPending() {
 
 export async function mempoolLookup(hash: string, chainId?: string) {
   let chainsToCheck = chainIds;
+  const { networkId, apiVersion } = await networkConfig;
 
   if (chainId) {
     chainsToCheck = [chainId];
@@ -63,7 +67,7 @@ export async function mempoolLookup(hash: string, chainId?: string) {
       const options = {
         hostname: dotenv.MEMPOOL_HOSTNAME,
         port: dotenv.MEMPOOL_PORT,
-        path: `/chainweb/0.0/development/chain/${chainId}/mempool/lookup`,
+        path: `/chainweb/${apiVersion}/${networkId}/chain/${chainId}/mempool/lookup`,
         method: 'POST',
         rejectUnauthorized: false, // This disables certificate verification
         headers: {