+++ packages/libs/react-ui/src/components/Button/utils.ts
@@ -2,14 +2,13 @@ import type { RecipeVariants } from '@vanilla-extract/recipes';
 import type { ComponentProps, ReactElement, ReactNode } from 'react';
 import { cloneElement } from 'react';
 import type { HoverEvents } from 'react-aria';
-import type { button } from './BaseButton/BaseButton.css';
-import { iconStyle } from './Button.css';
+import type { button } from './Button.css';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function disableLoadingProps<T>(props: T): T {
   const newProps: any = { ...props };
   // Don't allow interaction while isPending is true
-  if (newProps.isLoading) {
+  if (newProps?.isLoading) {
     newProps.onPress = undefined;
     newProps.onPressStart = undefined;
     newProps.onPressEnd = undefined;