+++ packages/libs/react-ui/src/components/NavHeader/NavHeader.stories.tsx
@@ -5,7 +5,6 @@ import {
 } from '@kadena/react-icons/system';
 import type { Meta, StoryObj } from '@storybook/react';
 import React, { useState } from 'react';
-import { tokens } from 'src/styles/tokens/contract.css';
 import { atoms } from '../../styles/atoms.css';
 import { SelectItem } from '../Form/Select';
 import { KadenaLogo } from '../Logo';
@@ -114,28 +113,10 @@ export const Dynamic: IStory = {
             </NavHeaderLink>
           ))}
         </NavHeaderLinkList>
+        <NavHeaderButton endVisual={<MonoAccountCircle />} />
         <NavHeaderButton
-          variant="transparent"
-          icon={
-            <MonoAccountCircle
-              className={atoms({
-                color: 'text.base.default',
-              })}
-            />
-          }
-        />
-        <NavHeaderButton
-          variant="transparent"
-          icon={
-            <MonoContrast
-              className={atoms({
-                color: 'text.base.default',
-              })}
-            />
-          }
-          className={atoms({
-            marginInlineEnd: 'sm',
-          })}
+          endVisual={<MonoContrast />}
+          className={atoms({ marginInlineEnd: 'sm' })}
         />
         <NavHeaderSelect
           aria-label="Select Network"