+++ packages/libs/react-ui/src/components/NavHeader/NavHeaderButton.tsx
@@ -2,6 +2,11 @@ import React from 'react';
 import type { IButtonProps } from '../Button';
 import { Button } from '../Button';
 
-export const NavHeaderButton = ({ className, ...props }: IButtonProps) => (
+export interface INavHeaderButtonProps extends IButtonProps {}
+
+export const NavHeaderButton = ({
+  className,
+  ...props
+}: INavHeaderButtonProps) => (
   <Button className={className} isCompact {...props} />
 );