+++ packages/apps/proof-of-us/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @kadena/proof-of-us
 
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [d82fbd97d]
+  - @kadena/graph@0.1.14
+
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [fe324038b]
+  - @kadena/graph@0.1.13
+
+## 0.2.7
+
+### Patch Changes
+
+- 8bb3168818: fix: use new text and header components from react-ui package
+- Updated dependencies [2bac09b5e]
+- Updated dependencies [b316674811]
+- Updated dependencies [8f7b1221c]
+- Updated dependencies [fa60f7fd9]
+- Updated dependencies [6aac4b623]
+- Updated dependencies [89412c156d]
+- Updated dependencies [2a7d0978ee]
+- Updated dependencies [3373d6397e]
+- Updated dependencies [0e5cceacc9]
+- Updated dependencies [e44ac6fdc]
+- Updated dependencies [fa2cce1cb]
+- Updated dependencies [f83bdeac1]
+- Updated dependencies [96c7b98341]
+- Updated dependencies [6514f74a5f]
+- Updated dependencies [7d1dbdd25]
+- Updated dependencies [b3af5a40e]
+- Updated dependencies [1ebfb0a52]
+- Updated dependencies [3bbfeaaa9]
+- Updated dependencies [3b2e0fe7c]
+- Updated dependencies [fc1308be85]
+- Updated dependencies [96e8c2556]
+- Updated dependencies [5a304c326]
+- Updated dependencies [6c2ec7fa9]
+- Updated dependencies [ffb089144]
+- Updated dependencies [fcdcf99f4]
+- Updated dependencies [8bb3168818]
+- Updated dependencies [a605cc1a6]
+- Updated dependencies [b7c5ca5789]
+- Updated dependencies [077c7c1e6]
+- Updated dependencies [5c5937eef6]
+  - @kadena/graph@0.1.12
+  - @kadena/react-ui@0.8.4
+  - @kadena/client@1.11.0
+
 ## 0.2.6
 
 ### Patch Changes